+++ backend/bots/examples/icp_dispenser/impl/src/lifecycle/post_upgrade.rs
@@ -1,11 +1,11 @@
-use crate::lifecycle::{init_logger, init_state, UPGRADE_BUFFER_SIZE};
-use crate::{Data, LOG_MESSAGES};
-use canister_logger::{LogMessage, LogMessagesWrapper};
+use crate::lifecycle::{init_state, UPGRADE_BUFFER_SIZE};
+use crate::Data;
 use canister_tracing_macros::trace;
 use ic_cdk_macros::post_upgrade;
 use icp_dispenser_bot::post_upgrade::Args;
 use stable_memory::deserialize_from_stable_memory;
 use tracing::info;
+use types::LogMessage;
 use utils::env::canister::CanisterEnv;
 
 #[post_upgrade]
@@ -18,26 +18,23 @@ fn post_upgrade(args: Args) {
     let (data, log_messages, trace_messages): (Data, Vec<LogMessage>, Vec<LogMessage>) =
         deserialize_from_stable_memory(UPGRADE_BUFFER_SIZE).unwrap();
 
-    init_logger(data.test_mode);
+    let logs = log_messages
+        .into_iter()
+        .map(|l| canister_logger::LogEntry {
+            timestamp: l.timestamp,
+            message: l.json,
+        })
+        .collect();
+    let traces = trace_messages
+        .into_iter()
+        .map(|t| canister_logger::LogEntry {
+            timestamp: t.timestamp,
+            message: t.json,
+        })
+        .collect();
+    canister_logger::init_with_logs(data.test_mode, logs, traces);
+
     init_state(env, data, args.wasm_version);
 
-    if !log_messages.is_empty() || !trace_messages.is_empty() {
-        LOG_MESSAGES.with(|l| rehydrate_log_messages(log_messages, trace_messages, &l.borrow()))
-    }
-
     info!(version = %args.wasm_version, "Post-upgrade complete");
-}
-
-fn rehydrate_log_messages(
-    log_messages: Vec<LogMessage>,
-    trace_messages: Vec<LogMessage>,
-    messages_container: &LogMessagesWrapper,
-) {
-    for message in log_messages {
-        messages_container.logs.push(message);
-    }
-
-    for message in trace_messages {
-        messages_container.traces.push(message);
-    }
 }