+++ backend/canisters/group_index/impl/src/updates/set_max_concurrent_group_canister_upgrades.rs
@@ -1,12 +1,12 @@
 use crate::guards::caller_is_controller;
 use crate::read_state;
+use canister_api_macros::proposal;
 use canister_tracing_macros::trace;
 use group_index_canister::set_max_concurrent_group_canister_upgrades::{Response::*, *};
-use ic_cdk_macros::update;
 use tracing::info;
 
 // dfx --identity openchat canister --network ic call group_index set_max_concurrent_group_canister_upgrades '(record { value=N:nat32 })'
-#[update(guard = "caller_is_controller")]
+#[proposal(guard = "caller_is_controller")]
 #[trace]
 async fn set_max_concurrent_group_canister_upgrades(args: Args) -> Response {
     let canisters: Vec<_> = read_state(|state| state.data.local_index_map.canisters().copied().collect());