+++ backend/canisters/user_index/api/src/updates/add_local_user_index_canister.rs
@@ -1,4 +1,5 @@
 use candid::CandidType;
+use human_readable::{HumanReadablePrincipal, ToHumanReadable};
 use serde::{Deserialize, Serialize};
 use types::CanisterId;
 
@@ -13,4 +14,21 @@ pub enum Response {
     Success,
     AlreadyAdded,
     InternalError(String),
+}
+
+#[derive(Serialize)]
+pub struct HumanReadableArgs {
+    canister_id: HumanReadablePrincipal,
+    notifications_canister_id: HumanReadablePrincipal,
+}
+
+impl ToHumanReadable for Args {
+    type Target = HumanReadableArgs;
+
+    fn to_human_readable(&self) -> Self::Target {
+        HumanReadableArgs {
+            canister_id: self.canister_id.into(),
+            notifications_canister_id: self.notifications_canister_id.into(),
+        }
+    }
 }