+++ backend/canisters/user_index/api/src/updates/add_super_admin.rs
@@ -1,4 +1,5 @@
-use candid::CandidType;
+use candid::{CandidType, Principal};
+use human_readable::{HumanReadablePrincipal, ToHumanReadable};
 use serde::{Deserialize, Serialize};
 use types::UserId;
 
@@ -12,4 +13,19 @@ pub enum Response {
     Success,
     AlreadySuperAdmin,
     InternalError(String),
+}
+
+#[derive(Serialize)]
+pub struct HumanReadableArgs {
+    user_id: HumanReadablePrincipal,
+}
+
+impl ToHumanReadable for Args {
+    type Target = HumanReadableArgs;
+
+    fn to_human_readable(&self) -> Self::Target {
+        HumanReadableArgs {
+            user_id: Principal::from(self.user_id).into(),
+        }
+    }
 }