+++ backend/canisters/user_index/api/src/updates/set_max_concurrent_user_canister_upgrades.rs
@@ -1,7 +1,8 @@
 use candid::CandidType;
+use human_readable::HumanReadable;
 use serde::{Deserialize, Serialize};
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(CandidType, Serialize, Deserialize, HumanReadable, Clone, Debug)]
 pub struct Args {
     pub value: u32,
 }