+++ backend/canisters/user_index/impl/src/updates/set_max_concurrent_user_canister_upgrades.rs
@@ -1,13 +1,13 @@
 use crate::guards::caller_is_controller;
 use crate::{mutate_state, RuntimeState};
+use canister_api_macros::proposal;
 use canister_tracing_macros::trace;
-use ic_cdk_macros::update;
 use local_user_index_canister::{Event, MaxConcurrentCanisterUpgradesChanged};
 use tracing::info;
 use user_index_canister::set_max_concurrent_user_canister_upgrades::{Response::*, *};
 
 // dfx --identity openchat canister --network ic call user_index set_max_concurrent_user_canister_upgrades '(record { value=N:nat32 })'
-#[update(guard = "caller_is_controller")]
+#[proposal(guard = "caller_is_controller")]
 #[trace]
 async fn set_max_concurrent_user_canister_upgrades(args: Args) -> Response {
     mutate_state(|state| set_max_concurrent_user_canister_upgrades_impl(args, state))