+++ backend/libraries/canister_client/src/operations/install_service_canisters.rs
@@ -1,80 +1,14 @@
-use crate::utils::{build_ic_agent, create_empty_canister, get_canister_wasm, install_wasm, set_controllers};
-use crate::{CanisterIds, CanisterName, CyclesDispenserInitArgs, OpenStorageCanisterName, OpenStorageInitArgs};
+use crate::utils::{build_ic_agent, get_canister_wasm, install_wasm, set_controllers};
+use crate::{
+    CanisterIds, CanisterName, CyclesDispenserConfig, CyclesDispenserInitArgs, OpenStorageCanisterName, OpenStorageInitArgs,
+};
 use candid::Principal;
 use ic_agent::identity::BasicIdentity;
 use ic_agent::{Agent, Identity};
 use ic_utils::interfaces::ManagementCanister;
-use types::{CanisterId, Version};
+use types::{Cycles, Version};
 
-const NNS_GOVERNANCE_CANISTER_ID: CanisterId = Principal::from_slice(&[0, 0, 0, 0, 0, 0, 0, 1, 1, 1]);
-
-pub async fn create_and_install_service_canisters(identity: BasicIdentity, url: String, test_mode: bool) -> CanisterIds {
-    let principal = identity.sender().unwrap();
-    let agent = build_ic_agent(url, identity).await;
-    let management_canister = ManagementCanister::create(&agent);
-
-    let (
-        user_index_canister_id,
-        group_index_canister_id,
-        notifications_index_canister_id,
-        local_user_index_canister_id,
-        local_group_index_canister_id,
-    ) = futures::future::join5(
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-    )
-    .await;
-
-    let (
-        notifications_canister_id,
-        online_users_canister_id,
-        proposals_bot_canister_id,
-        cycles_dispenser_canister_id,
-        open_storage_index_canister_id,
-    ) = futures::future::join5(
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-        create_empty_canister(&management_canister),
-    )
-    .await;
-
-    let ledger_canister_id = create_empty_canister(&management_canister).await;
-
-    println!("user_index canister id: {user_index_canister_id}");
-    println!("group_index canister id: {group_index_canister_id}");
-    println!("notifications index canister id: {notifications_index_canister_id}");
-    println!("local_user_index canister id: {local_user_index_canister_id}");
-    println!("local_group_index canister id: {local_group_index_canister_id}");
-    println!("notifications canister id: {notifications_canister_id}");
-    println!("online_users canister id: {online_users_canister_id}");
-    println!("proposals_bot canister id: {proposals_bot_canister_id}");
-    println!("cycles_dispenser canister id: {cycles_dispenser_canister_id}");
-    println!("open_storage_index canister id: {open_storage_index_canister_id}");
-    println!("ledger canister id: {ledger_canister_id}");
-
-    let canister_ids = CanisterIds {
-        user_index: user_index_canister_id,
-        group_index: group_index_canister_id,
-        notifications_index: notifications_canister_id,
-        local_user_index: local_user_index_canister_id,
-        local_group_index: local_group_index_canister_id,
-        notifications: notifications_canister_id,
-        online_users: online_users_canister_id,
-        proposals_bot: proposals_bot_canister_id,
-        cycles_dispenser: cycles_dispenser_canister_id,
-        open_storage_index: open_storage_index_canister_id,
-        ledger: ledger_canister_id,
-    };
-
-    install_service_canisters_impl(principal, &canister_ids, &agent, &management_canister, test_mode).await;
-
-    canister_ids
-}
+const T: Cycles = 1_000_000_000_000;
 
 pub async fn install_service_canisters(identity: BasicIdentity, url: String, canister_ids: CanisterIds, test_mode: bool) {
     let principal = identity.sender().unwrap();
@@ -131,7 +65,7 @@ async fn install_service_canisters_impl(
         group_index_canister_id: canister_ids.group_index,
         notifications_index_canister_id: canister_ids.notifications_index,
         open_storage_index_canister_id: canister_ids.open_storage_index,
-        ledger_canister_id: canister_ids.ledger,
+        ledger_canister_id: canister_ids.nns_ledger,
         proposals_bot_user_id: canister_ids.proposals_bot.into(),
         cycles_dispenser_canister_id: canister_ids.cycles_dispenser,
         wasm_version: version,
@@ -147,7 +81,7 @@ async fn install_service_canisters_impl(
         local_group_index_canister_wasm,
         user_index_canister_id: canister_ids.user_index,
         cycles_dispenser_canister_id: canister_ids.cycles_dispenser,
-        ledger_canister_id: canister_ids.ledger,
+        ledger_canister_id: canister_ids.nns_ledger,
         wasm_version: version,
         test_mode,
     };
@@ -178,7 +112,7 @@ async fn install_service_canisters_impl(
         service_owner_principals: vec![principal],
         user_index_canister_id: canister_ids.user_index,
         group_index_canister_id: canister_ids.group_index,
-        nns_governance_canister_id: NNS_GOVERNANCE_CANISTER_ID,
+        nns_governance_canister_id: canister_ids.nns_governance,
         cycles_dispenser_canister_id: canister_ids.cycles_dispenser,
         wasm_version: version,
         test_mode,
@@ -196,16 +130,24 @@ async fn install_service_canisters_impl(
             canister_ids.notifications,
             canister_ids.online_users,
             canister_ids.proposals_bot,
+            canister_ids.open_storage_index,
         ],
-        max_top_up_amount: 5_000_000_000_000,   // 5T
-        min_interval: 60 * 60 * 1000,           // 1 hour
-        min_cycles_balance: 10_000_000_000_000, // 10T
+        max_top_up_amount: 20 * T,
+        min_interval: 5 * 60 * 1000, // 5 minutes
+        min_cycles_balance: 200 * T,
+        icp_burn_amount_e8s: 1_000_000_000, // 10 ICP
+        ledger_canister: canister_ids.nns_ledger,
+        cycles_minting_canister: canister_ids.nns_cmc,
     };
 
     let open_storage_index_canister_wasm = get_canister_wasm(OpenStorageCanisterName::Index, version);
     let open_storage_index_init_args = OpenStorageInitArgs {
         service_principals: vec![principal, canister_ids.user_index, canister_ids.group_index],
         bucket_canister_wasm: get_canister_wasm(OpenStorageCanisterName::Bucket, version),
+        cycles_dispenser_config: Some(CyclesDispenserConfig {
+            canister_id: canister_ids.cycles_dispenser,
+            min_cycles_balance: 200 * T,
+        }),
         wasm_version: version,
         test_mode,
     };