+++ backend/libraries/types/src/lib.rs
@@ -115,3 +115,14 @@ impl<T: PartialEq> PushIfNotContains<T> for Vec<T> {
         }
     }
 }
+
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+
+// TODO remove this
+#[derive(CandidType, Serialize, Deserialize)]
+pub struct LogMessage {
+    pub timestamp: TimestampMillis,
+    #[serde(alias = "message")]
+    pub json: String,
+}